extract class field loading from member.ts into class-field-loading.ts
